scalar

+++ scalar/src/lib.rs
@@ -11,7 +11,7 @@ res
 
 }
 
-pub fn pro(a: i16, b: i16) -> i16 {
+pub fn pro(a: i8, b: i8) -> i8 {
 let  res;
 res = a*b;
 res
